ReColoring of particles and more seating configurations

+++ src/main/java/ar/edu/itba/ss2019b/model/Airplane.java
@@ -28,7 +28,7 @@ public class Airplane {
         return time;
     }
 
-    public double getPassengersSat(){
+    public int getPassengersSat(){
         return passengersSat;
     }
